+++ Elements/src/Geometry/Polyline.cs
@@ -10,7 +10,7 @@ namespace Elements.Geometry
     /// A coplanar continuous set of lines.
     /// </summary>
     /// <example>
-    /// [!code-csharp[Main](../../test/Elements.Tests/PolylineTests.cs?name=example)]
+    /// [!code-csharp[Main](../../Elements/test/PolylineTests.cs?name=example)]
     /// </example>
     public partial class Polyline : ICurve, IEquatable<Polyline>
     {